# Speed up playwright tests by 2x